+++ lib/glimmer/dsl/swt/cursor_expression.rb
@@ -1,0 +1,25 @@
+require 'glimmer/dsl/expression'
+require 'glimmer/dsl/top_level_expression'
+require 'glimmer/swt/cursor_proxy'
+
+module Glimmer
+  module DSL
+    module SWT
+      # cursor expression
+      # Note: Cannot be a static expression because it clashes with cursor property expression
+      class CursorExpression < Expression
+        include TopLevelExpression
+  
+        def can_interpret?(parent, keyword, *args, &block)          
+          (parent.nil? or !parent.respond_to?('cursor')) and
+            args.size == 1 and
+            (args.first.is_a?(Integer) or textual?(args.first))
+        end
+  
+        def interpret(parent, keyword, *args, &block)
+          Glimmer::SWT::CursorProxy.new(*args)
+        end
+      end
+    end
+  end
+end